move setSignals to cli/js/signals.ts
